refactor: decouple scripting environment imports
