+++ packages/insomnia-data/common-src/index.ts
@@ -4,6 +4,10 @@ export * from './ndjson';
 export * from './type';
 export * from './misc';
 export * from './constants';
+export * from './response-reasons';
+export * from './faker-functions';
+export * from './protocol';
+export * from './url-matches-cert-host';
 export * from './preview-mode';
 export * from './platform';
 export * from './strings';
