+++ packages/insomnia-scripting-environment/src/objects/interpolator.ts
@@ -1,4 +1,4 @@
-import { fakerFunctions } from 'insomnia/src/common/templating/faker-functions';
+import { fakerFunctions } from 'insomnia-data/common';
 import { Liquid } from 'liquidjs';
 
 /** @ignore */
