+++ packages/insomnia-scripting-environment/vitest.config.ts
@@ -3,8 +3,5 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     hideSkippedTests: true,
-    alias: {
-      '~/': new URL('../insomnia/src/', import.meta.url).pathname,
-    },
   },
 });
