+++ packages/insomnia/src/common/templating/local-template-tags.ts
@@ -1,13 +1,12 @@
 import { format } from 'date-fns';
 import type { PluginTemplateTag } from 'insomnia/src/common/templating/types';
+import { fakerFunctions } from 'insomnia-data/common';
 import JSONBig from 'json-bigint';
 import { JSONPath } from 'jsonpath-plus';
 
 import type { TemplateTag } from '~/common/plugins/types';
 import { invariant } from '~/common/utils/invariant';
 import { utf8StringFromBytes } from '~/common/utils/utf8-bytes';
-
-import { fakerFunctions } from './faker-functions';
 
 const JSONBigStringParser = JSONBig({ storeAsString: true });
