+++ packages/insomnia/src/plugins/context/network.ts
@@ -1,9 +1,9 @@
 import { services } from 'insomnia-data';
+import { RESPONSE_CODE_REASONS } from 'insomnia-data/common';
 import { v4 as uuidv4 } from 'uuid';
 
 import type { NodeCurlRequestOptions, NodeCurlResponseType, PluginTemplateTagContext } from '~/common/templating/types';
 
-import { RESPONSE_CODE_REASONS } from '../../common/constants';
 import {
   fetchRequestData,
   responseTransform,
